requirements.txt
fastapi
uvicorn[standard]
python-dotenv
torch
openai-whisper
python-multipart
groq
gtts
requests
tenacity
pandas